+++ lib/chewie_audio.dart
@@ -2,3 +2,5 @@ library chewie_audio;
 
 export 'src/chewie_player.dart';
 export 'src/chewie_progress_colors.dart';
+export 'src/cupertino_controls.dart';
+export 'src/material_controls.dart';