+++ lib/src/material_controls.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:chewie_audio/src/animated_play_pause.dart';
 import 'package:chewie_audio/src/chewie_player.dart';
 import 'package:chewie_audio/src/chewie_progress_colors.dart';
 import 'package:chewie_audio/src/material_progress_bar.dart';
@@ -16,8 +17,9 @@ class MaterialControls extends StatefulWidget {
   }
 }
 
-class _MaterialControlsState extends State<MaterialControls> with SingleTickerProviderStateMixin {
-  VideoPlayerValue? _latestValue;
+class _MaterialControlsState extends State<MaterialControls>
+    with SingleTickerProviderStateMixin {
+  late VideoPlayerValue _latestValue;
   double? _latestVolume;
   Timer? _initTimer;
   Timer? _showAfterExpandCollapseTimer;
@@ -25,28 +27,39 @@ class _MaterialControlsState extends State<MaterialControls> with SingleTickerPr
   final barHeight = 48.0;
   final marginSize = 5.0;
 
-  VideoPlayerController? controller;
-  ChewieAudioController? chewieController;
-  AnimationController? playPauseIconAnimationController;
+  late VideoPlayerController controller;
+  ChewieAudioController? _chewieController;
+  // We know that _chewieController is set in didChangeDependencies
+  ChewieAudioController get chewieController => _chewieController!;
 
   @override
   Widget build(BuildContext context) {
-    if (_latestValue!.hasError) {
-      return chewieController!.errorBuilder != null
-          ? chewieController!.errorBuilder!(
-              context,
-              chewieController!.videoPlayerController.value.errorDescription,
-            )
-          : const Center(
-              child: Icon(
-                Icons.error,
-                color: Colors.white,
-                size: 42,
-              ),
-            );
+    if (_latestValue.hasError) {
+      return chewieController.errorBuilder?.call(
+            context,
+            chewieController.videoPlayerController.value.errorDescription!,
+          ) ??
+          const Center(
+            child: Icon(
+              Icons.error,
+              color: Colors.white,
+              size: 42,
+            ),
+          );
     }
 
-    return _buildBottomBar(context);
+    return Column(
+      children: <Widget>[
+        if (_latestValue.isBuffering)
+          const Expanded(
+            child: Center(
+              child: CircularProgressIndicator(),
+            ),
+          )
+        else
+        _buildBottomBar(context),
+      ],
+    );
   }
 
   @override
@@ -56,22 +69,16 @@ class _MaterialControlsState extends State<MaterialControls> with SingleTickerPr
   }
 
   void _dispose() {
-    controller!.removeListener(_updateState);
+    controller.removeListener(_updateState);
     _initTimer?.cancel();
     _showAfterExpandCollapseTimer?.cancel();
   }
 
   @override
   void didChangeDependencies() {
-    final _oldController = chewieController;
-    chewieController = ChewieAudioController.of(context);
-    controller = chewieController!.videoPlayerController;
-
-    playPauseIconAnimationController ??= AnimationController(
-      vsync: this,
-      duration: const Duration(milliseconds: 400),
-      reverseDuration: const Duration(milliseconds: 400),
-    );
+    final _oldController = _chewieController;
+    _chewieController = ChewieAudioController.of(context);
+    controller = chewieController.videoPlayerController;
 
     if (_oldController != chewieController) {
       _dispose();
@@ -81,7 +88,7 @@ class _MaterialControlsState extends State<MaterialControls> with SingleTickerPr
     super.didChangeDependencies();
   }
 
-  Widget _buildBottomBar(
+  Container _buildBottomBar(
     BuildContext context,
   ) {
     final iconColor = Theme.of(context).textTheme.button!.color;
@@ -91,24 +98,64 @@ class _MaterialControlsState extends State<MaterialControls> with SingleTickerPr
       color: Theme.of(context).dialogBackgroundColor,
       child: Row(
         children: <Widget>[
-          _buildPlayPause(controller!),
-          if (chewieController!.isLive) const Expanded(child: Text('LIVE')) else _buildPosition(iconColor),
-          if (chewieController!.isLive) const SizedBox() else _buildProgressBar(),
-          if (chewieController!.allowMuting) _buildMuteButton(controller),
+          _buildPlayPause(controller),
+          if (chewieController.isLive)
+            const Expanded(child: Text('LIVE'))
+          else
+            _buildPosition(iconColor),
+          if (chewieController.isLive)
+            const SizedBox()
+          else
+            _buildProgressBar(),
+          if (chewieController.allowPlaybackSpeedChanging)
+            _buildSpeedButton(controller),
+          if (chewieController.allowMuting) _buildMuteButton(controller),
         ],
       ),
     );
   }
 
+  Widget _buildSpeedButton(
+    VideoPlayerController controller,
+  ) {
+    return GestureDetector(
+      onTap: () async {
+        final chosenSpeed = await showModalBottomSheet<double>(
+          context: context,
+          isScrollControlled: true,
+          useRootNavigator: true,
+          builder: (context) => _PlaybackSpeedDialog(
+            speeds: chewieController.playbackSpeeds,
+            selected: _latestValue.playbackSpeed,
+          ),
+        );
+
+        if (chosenSpeed != null) {
+          controller.setPlaybackSpeed(chosenSpeed);
+        }
+      },
+      child: ClipRect(
+        child: Container(
+          height: barHeight,
+          padding: const EdgeInsets.only(
+            left: 8.0,
+            right: 8.0,
+          ),
+          child: const Icon(Icons.speed),
+        ),
+      ),
+    );
+  }
+
   GestureDetector _buildMuteButton(
-    VideoPlayerController? controller,
+    VideoPlayerController controller,
   ) {
     return GestureDetector(
       onTap: () {
-        if (_latestValue!.volume == 0) {
-          controller!.setVolume(_latestVolume ?? 0.5);
+        if (_latestValue.volume == 0) {
+          controller.setVolume(_latestVolume ?? 0.5);
         } else {
-          _latestVolume = controller!.value.volume;
+          _latestVolume = controller.value.volume;
           controller.setVolume(0.0);
         }
       },
@@ -120,7 +167,7 @@ class _MaterialControlsState extends State<MaterialControls> with SingleTickerPr
             right: 8.0,
           ),
           child: Icon(
-            (_latestValue != null && _latestValue!.volume > 0) ? Icons.volume_up : Icons.volume_off,
+            _latestValue.volume > 0 ? Icons.volume_up : Icons.volume_off,
           ),
         ),
       ),
@@ -138,16 +185,16 @@ class _MaterialControlsState extends State<MaterialControls> with SingleTickerPr
           left: 12.0,
           right: 12.0,
         ),
-        child: Icon(
-          controller.value.isPlaying ? Icons.pause : Icons.play_arrow,
+        child: AnimatedPlayPause(
+          playing: controller.value.isPlaying,
         ),
       ),
     );
   }
 
   Widget _buildPosition(Color? iconColor) {
-    final position = _latestValue != null && _latestValue!.position != null ? _latestValue!.position : Duration.zero;
-    final duration = _latestValue != null && _latestValue!.duration != null ? _latestValue!.duration : Duration.zero;
+    final position = _latestValue.position;
+    final duration = _latestValue.duration;
 
     return Padding(
       padding: const EdgeInsets.only(right: 24.0),
@@ -161,35 +208,27 @@ class _MaterialControlsState extends State<MaterialControls> with SingleTickerPr
   }
 
   Future<void> _initialize() async {
-    controller!.addListener(_updateState);
+    controller.addListener(_updateState);
 
     _updateState();
   }
 
   void _playPause() {
-    bool isFinished;
-    if (_latestValue!.duration != null) {
-      isFinished = _latestValue!.position >= _latestValue!.duration;
-    } else {
-      isFinished = false;
-    }
+    final isFinished = _latestValue.position >= _latestValue.duration;
 
     setState(() {
-      if (controller!.value.isPlaying) {
-        playPauseIconAnimationController!.reverse();
-        controller!.pause();
+      if (controller.value.isPlaying) {
+        controller.pause();
       } else {
-        if (!controller!.value.isInitialized) {
-          controller!.initialize().then((_) {
-            controller!.play();
-            playPauseIconAnimationController!.forward();
+        if (!controller.value.isInitialized) {
+          controller.initialize().then((_) {
+            controller.play();
           });
         } else {
           if (isFinished) {
-            controller!.seekTo(const Duration());
+            controller.seekTo(const Duration());
           }
-          playPauseIconAnimationController!.forward();
-          controller!.play();
+          controller.play();
         }
       }
     });
@@ -197,7 +236,7 @@ class _MaterialControlsState extends State<MaterialControls> with SingleTickerPr
 
   void _updateState() {
     setState(() {
-      _latestValue = controller!.value;
+      _latestValue = controller.value;
     });
   }
 
@@ -209,7 +248,7 @@ class _MaterialControlsState extends State<MaterialControls> with SingleTickerPr
           controller,
           onDragStart: () {},
           onDragEnd: () {},
-          colors: chewieController!.materialProgressColors ??
+          colors: chewieController.materialProgressColors ??
               ChewieProgressColors(
                   playedColor: Theme.of(context).accentColor,
                   handleColor: Theme.of(context).accentColor,
@@ -217,6 +256,54 @@ class _MaterialControlsState extends State<MaterialControls> with SingleTickerPr
                   backgroundColor: Theme.of(context).disabledColor),
         ),
       ),
+    );
+  }
+}
+
+class _PlaybackSpeedDialog extends StatelessWidget {
+  const _PlaybackSpeedDialog({
+    Key? key,
+    required List<double> speeds,
+    required double selected,
+  })   : _speeds = speeds,
+        _selected = selected,
+        super(key: key);
+
+  final List<double> _speeds;
+  final double _selected;
+
+  @override
+  Widget build(BuildContext context) {
+    final Color selectedColor = Theme.of(context).primaryColor;
+
+    return ListView.builder(
+      shrinkWrap: true,
+      physics: const ScrollPhysics(),
+      itemBuilder: (context, index) {
+        final _speed = _speeds[index];
+        return ListTile(
+          dense: true,
+          title: Row(
+            children: [
+              if (_speed == _selected)
+                Icon(
+                  Icons.check,
+                  size: 20.0,
+                  color: selectedColor,
+                )
+              else
+                Container(width: 20.0),
+              const SizedBox(width: 16.0),
+              Text(_speed.toString()),
+            ],
+          ),
+          selected: _speed == _selected,
+          onTap: () {
+            Navigator.of(context).pop(_speed);
+          },
+        );
+      },
+      itemCount: _speeds.length,
     );
   }
 }